+++ polymer-perf/eduke32/source/savegame.c
@@ -937,7 +937,7 @@ static uint8_t savegame_comprthres;
 // 64
 #define DS_LOADFN 128  // .ptr is function that is run when loading
 #define DS_SAVEFN 256  // .ptr is function that is run when saving
-#define DS_NOCHK 1024  // don't check for diffs (and don't write out in dump) since assumend constant throughout demo
+#define DS_NOCHK 1024  // don't check for diffs (and don't write out in dump) since assumed constant throughout demo
 #define DS_END (0x70000000)
 
 static int32_t ds_getcnt(const dataspec_t *sp)
@@ -1721,7 +1721,7 @@ int32_t sv_loadsnapshot(int32_t fil, int32_t *ret_hasdiffs, int32_t *ret_demotic
     if (p != svsnapshot+svsnapsiz)
     {
         OSD_Printf("sv_loadsnapshot: internal error: p-(snapshot end)=%d!\n",
-                   p-(svsnapshot+svsnapsiz));
+                   (int32_t)(p-(svsnapshot+svsnapsiz)));
         sv_freemem();
         return 6;
     }
@@ -2042,7 +2042,7 @@ static void sv_restload()
 #define SAVEWR(ptr, sz, cnt) do { if (fil) dfwrite(ptr,sz,cnt,fil); } while (0)
 #define SAVEWRU(ptr, sz, cnt) do { if (fil) fwrite(ptr,sz,cnt,fil); } while (0)
 
-#define PRINTSIZE(name) OSD_Printf(#name ": %d\n", mem-tmem), tmem=mem
+#define PRINTSIZE(name) OSD_Printf(#name ": %d\n", (int32_t)(mem-tmem)), tmem=mem
 
 static uint8_t *dosaveplayer2(int32_t spot, FILE *fil, uint8_t *mem)
 {
